Add files via upload
This piece of code implements the Bruun Rule from within ArcMap. The Bruun Rule is a formula for estimating the magnitude of the retreat of the shoreline of a sandy shore in response to changes in sea level. The Bruun Rule gives a linear relationship between sea level rise and shoreline recession based on equilibrium profile theory, which asserts that shore face profile maintains an equilibrium shape, and as sea level rises the increasing accommodation space forces this equilibrium profile landward and upward to preserve its shape relative to the new sea level. As such, the Bruun Rule analysis assumes that the upper beach is eroded as the shore profile moves landward, and that the volume of eroded material is deposited offshore, resulting in a rise of the nearshore bottom which maintains a constant water depth