Add files via upload
Simulating crime occurrence using Artificially Intelligent agents. Re-inforcement learning is used to ensure the agents learn from experience whereas agent-based modeling in netlogo models the human behaviour into the agents, GIS extension was used to build a risk terrain model which is the actual space that the interraction of the agents occurs. Crime pattern is observed by way of time-step graphs and simulated crime occurrencces are mapped using the GIS extension to netlogo.